Migrated database.py to class, get function now returns Database object
The database class previously only returned the formatted URL string and not the object itself requiring us to import dataset in any file where a database operation was occurring which was redundant. 

get_db() was renamed to get() and now returns the Dataset database object. The new url() function has the same functionality as the old get_db() which is used for creating the database if it didn't already exist.