Removed before_invoke, has_permissions, and staticmethod decorators
Due to the way that cross-file function calls are done in discord.py, the usages of @staticmethod were useless bloat.

Since the switch to discord-py-slash-commands/discord-interactions, before_invoke and has_permissions have been broken. 

has_permission has more or less been replaced by the new permissions system for slash commands but we need to find a way to re-implement before_invoke in the future to log user commands to console.