Moved cog loading into on_ready and made the loading more dynamic
The previous implementation made it so that cogs could not access any bot values during their initialization because the bot was loading the cogs before connecting to the Discord network. This fixes that going forward.

The root_dir for the cog loading was changed so that it will uses the parent directory of the executing script rather than just using a hardcoded "chiya" value which is rather unnecessarily static.

Removed the legacy comment about registering commands because this seems to be fixed in later versions of Pycord although we plan to switch to discord.py eventually anyways.