docker-publish.yml now properly builds to repo
- Fixed name so Docker images are now located under snaacky/chiya:latest instead of snaacky/chiya/chiya:latest.
- Implemented caching to hopefully speed up Docker image builds (cache is located under chiya-stages which is hidden from non-maintainers).
- Simplified bloat in docker-publish.yml because there was a lot of lines that were never used.
- Removed the "test" job which was just a Docker image build that never got pushed anywhere before the actual build occurred because it just increased total build time for no reason.

It seems like performance for this is all over the board, seems mostly out of our control, but I've seen compile times as low as 50s now which seems to be a net gain. We can probably tweak this more as time goes on if it needs extra time in the oven.

Using https://github.com/whoan/docker-build-with-cache-action for the caching action.