Revamp Boosting improve #31
- Mentions member who boosted in message.
- Moved log message to same function as general message.
- Added jump link to log boost embed.

I was going to change the message from messaging staff to creating a ticket and mentioning the ticket channel but we don't have config for that. So if add that I can change it to mention the ticket channel.

I tested the message.author on a joining guild message and it returns the member joining so I assume it works the same for boosting