Fixed database collation check so it doesn't run every bot start
Previously, we were just setting the collation every single time the bot ran which was completely unnecessary because it should never change after the database is initially created. This check will verify the character set at the cost of a few extra queries (which should theoretically cause 0% noticeable impact) and only update the collation when necessary. Annoyingly, dataset does not expose the collation in any capacity despite it being exposed by the underlying SQLAlchemy.